Fix escaping for perfmon metrics export (#121)
Fix escaping for perfmon metrics export

What occurs?
Whenever new log rows, containing database names with some special characters,
were tried to be uploaded from _queries_history.dat file to the
queries_history external table an error considering the size of a db column
would occur. It would prevent any new logs to be loaded whatsoever.

Why it occurs?
The db column has a type of VARCHAR(64), thus it means that a larger string
was tried to be put. Obvious reason for this - incorrect escaping or lack of
such whatsoever. Only two symbols were observed to lead to errors: " and |.
In case of a pipe - it leads to another error about incorrect row structure,
which is logical, as pipes are used as delimiters inside the file.

How do we fix it?
Before the patch (1f67d39) db field used to be written to
_queries_history.dat without double quotes, so no escaping was present. Now
it has it, as we enclose every database name in double quotes. Also, we double
the already present in database name double quotes.

Does it fix the problem?
Yes, as now we escape the whole database name string with all of its special
(or not) characters - the very same method is used for the SQL query command
and it works. Doubling double quotes is needed as we need to escape the quotes
to not to end the string to early.

What was changed?
Minor code additions for escaping inside the src/backend/gpmon/gpmon.c

Tests?
New BDD auto test to check that before mentioned logs are added to
queries_history correctly (as they appear there). But, not all symbols can be
auto tested using present testing functions: ', | and UTF-8 symbols. Errors
occur at different steps of a test for different type of symbol and are
connected to the way commands are issued to the DBSM. Nevertheless, during hand
testing these symbols passed.

Observations?
dbuser column also is not escaped, but I have not managed to recreate the
same issue with it. Yet, it may be worth to add escaping to it in future, but
now it seems like an extremely edge case scenario.

Ticket: ADBDEV-7272

---------

Co-authored-by: Vladimir Sarmin <v.sarmin@arenadata.io>
Co-authored-by: Georgy Shelkovy <g.shelkovy@arenadata.io>