Fix inconsistency DROP IF EXISTS command (#130)
When the DROP IF EXISTS command is called, it is unconditionally dispatched to
the segments, even if the object doesn't exist on the coordinator. This can
lead to inconsistencies, for example:

- the first session starts a table drop on the coordinator. Even though the
table doesn't exist, the drop is still dispatched to the segments.

- at the same time, the second session creates the table (on the coordinator
and on the segments).

- the first session (on the segments) already sees the table and therefore
deletes it.

Don't dispatch DROP if object doesn't exist on coordinator. Exclude the
drop_rename test from the parallel group because it contains fault-injectors.
Add a new GUC, gp_dispatch_drop_always, for unconditional DROP dispatch even if
the object isn't present on the coordinator, as this functionality is used by
the gpcheckcat utility, for example, in the orphan_temp_table test.

Ticket: ADBDEV-8867