Fix vacuum doesn't compact AO RO table after rolled back transaction (#151)
Problem description:
In case a column was added to a AO row-oriented table in a transaction (and
some values were set to this new column), and this transaction was rolled back,
vacuum didn't compact the table.

Root cause:
In case a new column is added to a AO row-oriented table, the table has to be
rewritten, and a new 'pg_aoseg_...' table is created. Segment data from the new
segment table is also stored to 'AppendOnlyHash' hash table entry.  If we
rollback the transaction, 'pg_aoseg_...' table is rolled back to the old one. 
But the hash entry continued to exist with the values from the transient
pg_aoseg table. So, during vacuum we tried to get the list of segments we had
to compact and called SetSegnoForCompaction(), which requested data from the
hash (via AORelGetOrCreateHashEntry()) and got the hash entry with 0 tupcount.
Therefore, compaction didn't happen.

Fix:
Remove the hash entry from 'AppendOnlyHash' in 'AtEOXact_AppendOnly_Relation()'
if the transaction is aborted. It is done for both row and column oriented
tables, as both are susceptible to the non-valid hash entry.

'AtEOXact_AppendOnly_Relation()' previously contained the code to remove the
hash entry, but it was under condition of GUC
'test_AppendOnlyHash_eviction_vs_just_marking_not_inuse'. This GUC did nothing
more, and was disabled by default. So, as this patch adds proper handling for
the hash entry eviction, it also makes this GUC deprecated and removes its
usage from 'ao-insert-eof' isolation test.